backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
redis
numpy
scikit-learn
faiss-cpu
sentence-transformers
numpy
requests